Make pools remove their own stale jobs in the event of a hard failure.
Since each worker pool now has a unique hardware id, we can now purge
the stale jobs for a pool in the event of a hard failure and reboot.
When the pool restarts, the first thing it does is remove jobs that
belong to its own id that are in the executing set (if any) and
re-queue them. This skips the unecessary step of having a pool attempt
to ping itself.