Use scripts/generate.go to generate a scripts.go file that reads strings instead of files.
This makes jobs suitable for use in executables without needing to include the scripts folder.
It also makes jobs safe for use with any package manager of GOPATH setup. The onus is on me
(and other contributors) to rememver to run go generate . after changing any of the lua scripts.