Export all fields of Job via methods. Fixes #16.
The fact that the fields are exported via getter methods means
the caller can get the current value but can't change it. This
is the desired behavior because the current architecture doesn't
let you change certain properties of the job directly. (You need
to do so with the `Reschedule` method, for example, because that
method makes sure the change is atomic).